Refactor: Cardano node configuration in infra

+++ mithril-infra/mithril.aggregator.tf
@@ -9,10 +9,19 @@ resource "null_resource" "mithril_aggregator" {
   ]
 
   triggers = {
-    image_id                         = var.mithril_image_id,
-    vm_instance                      = google_compute_instance.vm_instance.id,
-    mithril_aggregator_auth_username = var.mithril_aggregator_auth_username,
-    mithril_aggregator_auth_password = var.mithril_aggregator_auth_password
+    vm_instance                                                           = google_compute_instance.vm_instance.id,
+    cardano_image_id                                                      = var.cardano_image_id,
+    cardano_image_registry                                                = var.cardano_image_registry,
+    image_id                                                              = var.mithril_image_id,
+    mithril_aggregator_auth_username                                      = var.mithril_aggregator_auth_username,
+    mithril_aggregator_auth_password                                      = var.mithril_aggregator_auth_password
+    mithril_aggregator_signed_entity_types                                = var.mithril_aggregator_signed_entity_types,
+    mithril_aggregator_snapshot_compression_algorithm                     = var.mithril_aggregator_snapshot_compression_algorithm,
+    mithril_aggregator_zstandard_parameters_level                         = var.mithril_aggregator_zstandard_parameters_level,
+    mithril_aggregator_zstandard_parameters_workers                       = var.mithril_aggregator_zstandard_parameters_workers,
+    mithril_aggregator_cardano_transactions_prover_cache_pool_size        = var.mithril_aggregator_cardano_transactions_prover_cache_pool_size,
+    mithril_aggregator_cardano_transactions_database_connection_pool_size = var.mithril_aggregator_cardano_transactions_database_connection_pool_size,
+    mithril_aggregator_cexplorer_pools_url                                = var.mithril_aggregator_cexplorer_pools_url,
   }
 
   connection {
@@ -33,11 +42,30 @@ resource "null_resource" "mithril_aggregator" {
       <<-EOT
 set -e
 # Setup cardano node configuration
-AGGREGATOR_CONFIG_DIRECTORY=/home/curry/data/${var.cardano_network}/mithril-aggregator/cardano/config
-cp -R /home/curry/docker/cardano-configurations/network/${var.cardano_network} $AGGREGATOR_CONFIG_DIRECTORY
-cat $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json | jq ".hasPrometheus[0] |= \"cardano-node-aggregator\"" > $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json.new
-rm -f $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json
-mv $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json.new $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json
+# Copy the cardano node configuration files to the aggregator (exact version, and fallback to minor version)
+CARDANO_NODE_EXACT_VERSION="${var.cardano_image_id}"
+CARDANO_NODE_MINOR_VERSION=$(echo $CARDANO_NODE_EXACT_VERSION | cut -d. -f1,2)
+CARDANO_NODE_VERSIONS="$CARDANO_NODE_EXACT_VERSION $CARDANO_NODE_MINOR_VERSION"
+FOUND_CONFIGURATION=false
+for CARDANO_NODE_VERSION in $CARDANO_NODE_VERSIONS; do
+  if [ -d "/home/curry/docker/cardano/config/$CARDANO_NODE_VERSION/${var.cardano_network}" ]; then 
+    AGGREGATOR_CONFIG_DIRECTORY=/home/curry/data/${var.cardano_network}/mithril-aggregator/cardano/config
+    rm -rf $AGGREGATOR_CONFIG_DIRECTORY
+    mkdir -p $AGGREGATOR_CONFIG_DIRECTORY
+    cp -R /home/curry/docker/cardano/config/$CARDANO_NODE_VERSION/${var.cardano_network} $AGGREGATOR_CONFIG_DIRECTORY
+    echo $CARDANO_NODE_VERSION > $AGGREGATOR_CONFIG_DIRECTORY/config.version
+    cat $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json | jq ".hasPrometheus[0] |= \"cardano-node-aggregator\"" > $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json.new
+    rm -f $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json
+    mv $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json.new $AGGREGATOR_CONFIG_DIRECTORY/${var.cardano_network}/cardano-node/config.json
+    FOUND_CONFIGURATION=true
+    break
+  fi
+done
+# Check if a configuration was found
+if [ "$FOUND_CONFIGURATION" = "false" ]; then
+  echo "No cardano node configuration found for version $CARDANO_NODE_EXACT_VERSION"
+  exit 1
+fi
 EOT
     ]
   }
